fix: remove unnecessary cast in getDamageModifier method in Inventory class

+++ java/src/main/java/codingdojo/Inventory.java
@@ -12,7 +12,7 @@ public class Inventory {
     }
 
     public float getDamageModifier(){
-        return (float) equipment.getDamageModifier();
+        return equipment.getDamageModifier();
     }
     }
